Add WFQ (to Policy.t) and bin/sim.ml

+++ rio/simulator/rank.ml
@@ -1,1 +1,1 @@
-type t = int
+type t = float
